Add files via upload
Login system working, however, it allows the user to use the same email to register twice but the second registered email can not login using the same email. The search function is complete, however, the theme background did not work on the search page and user information page, lastly it can not recognise username if data entered backwards ( maybe better like this). Later, will look into Ajax loading and SEO if needed.